Add role to build DNNL
This role adds the building of DNNL (or mkl-dnn in this version)
to the playbook.
DNNL will be used by TensorFlow to optimize via JIT the kernels
for the specific microarchitecture (and esp. for SVE).
Note that this commit does not hook the built DNNL to TensorFlow,
since it requires some Bazel build files investigation.